requirements.txt
Flask>=3.0
Pillow>=10.0
rembg[cpu]>=2.0
numpy>=1.24
opencv-python>=4.8
